requirements.txt
# Python dependencies for CUB-200-2011 Knowledge Distillation

# Model libraries
transformers>=4.30.0
timm>=0.9.0
qwen-vl-utils  # For Qwen2.5-VL support
accelerate>=0.20.0  # For model loading optimization

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development (optional)
jupyter>=1.0.0
ipython>=8.12.0
